Change Calendly buttons to a standard link for AMP requests
AMP strips out `onClick` handlers, so when the button version of the
block is rendered we should change it to a standard link to the Calendly
appointment page.

We could probably get aronud this by attaching the click handler in
another way, but it would still be good to treat an AMP request
differently.